fix: billing records
